Implementação do Card de Visitantes com Detalhamento e Opções de Gerenciamento
Adicionado Card de Visitantes que lista todos os visitantes cadastrados.
Permite a seleção de um visitante para exibir seus detalhes no Sidebar.
Detalhes do Visitante incluem:
Nome, Documento, Empresa.
Lista de visitas relacionadas, exibida em formato tabular.
Adicionado:
Botão para excluir o visitante diretamente do Sidebar.
Botão ao lado de cada linha de visita para excluir visitas individuais.
Otimizada a experiência visual e interativa:
Coluna de "Ações" na tabela é exibida apenas quando há visitas registradas.
Inclusão de transições para melhor usabilidade do Sidebar.
Preparado para exportação:
Botão para baixar os detalhes do visitante (dados pessoais e visitas) em um arquivo Excel.

+++ src/main/java/com/servicos/cadastro_servicos/service/CategoriaService.java
@@ -29,6 +29,11 @@ public class CategoriaService {
     }
 
     public void deletar(Long id) {
-        categoriaRepository.deleteById(id);
+        if (categoriaRepository.existsById(id)) {
+            categoriaRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Categoria com ID " + id + " não encontrada.");
+        }
     }
+    
 }
